backend/requirements.txt
fastapi
uvicorn
pydantic
gitpython
pytest
langgraph
langchain
langchain-openai
python-dotenv
